Split metadata parsing
In preparation to vendoring Metadata23 from https://github.com/PyO3/python-pkginfo-rs/blob/main/src/metadata.rs for https://github.com/astral-sh/uv/pull/7475#pullrequestreview-2323104398
