+++ crates/uv-cache/src/lib.rs
@@ -11,7 +11,7 @@ use tracing::debug;
 
 pub use archive::ArchiveId;
 use distribution_types::InstalledDist;
-use pypi_types::Metadata23;
+use pypi_types::MetadataResolver;
 use uv_cache_info::Timestamp;
 use uv_fs::{cachedir, directories};
 use uv_normalize::PackageName;
@@ -789,7 +789,7 @@ impl CacheBucket {
             let Ok(metadata) = fs_err::read(path.join("metadata.msgpack")) else {
                 return false;
             };
-            let Ok(metadata) = rmp_serde::from_slice::<Metadata23>(&metadata) else {
+            let Ok(metadata) = rmp_serde::from_slice::<MetadataResolver>(&metadata) else {
                 return false;
             };
             metadata.name == *name
