+++ crates/uv-distribution/src/download.rs
@@ -3,7 +3,7 @@ use std::path::{Path, PathBuf};
 use crate::Error;
 use distribution_filename::WheelFilename;
 use distribution_types::{CachedDist, Dist, Hashed};
-use pypi_types::{HashDigest, Metadata23};
+use pypi_types::{HashDigest, MetadataResolver};
 use uv_metadata::read_flat_wheel_metadata;
 
 use uv_cache_info::CacheInfo;
@@ -40,8 +40,8 @@ impl LocalWheel {
         &self.filename
     }
 
-    /// Read the [`Metadata23`] from a wheel.
-    pub fn metadata(&self) -> Result<Metadata23, Error> {
+    /// Read the [`MetadataResolver`] from a wheel.
+    pub fn metadata(&self) -> Result<MetadataResolver, Error> {
         read_flat_wheel_metadata(&self.filename, &self.archive)
             .map_err(|err| Error::WheelMetadata(self.archive.clone(), Box::new(err)))
     }
