+++ crates/uv-distribution/src/metadata/mod.rs
@@ -4,7 +4,7 @@ use std::path::Path;
 use thiserror::Error;
 
 use pep440_rs::{Version, VersionSpecifiers};
-use pypi_types::{HashDigest, Metadata23};
+use pypi_types::{HashDigest, MetadataResolver};
 use uv_configuration::SourceStrategy;
 use uv_normalize::{ExtraName, GroupName, PackageName};
 use uv_workspace::WorkspaceError;
@@ -39,7 +39,7 @@ pub struct Metadata {
 impl Metadata {
     /// Lower without considering `tool.uv` in `pyproject.toml`, used for index and other archive
     /// dependencies.
-    pub fn from_metadata23(metadata: Metadata23) -> Self {
+    pub fn from_metadata23(metadata: MetadataResolver) -> Self {
         Self {
             name: metadata.name,
             version: metadata.version,
@@ -57,7 +57,7 @@ impl Metadata {
     /// Lower by considering `tool.uv` in `pyproject.toml` if present, used for Git and directory
     /// dependencies.
     pub async fn from_workspace(
-        metadata: Metadata23,
+        metadata: MetadataResolver,
         install_path: &Path,
         sources: SourceStrategy,
     ) -> Result<Self, MetadataError> {
@@ -102,7 +102,7 @@ pub struct ArchiveMetadata {
 impl ArchiveMetadata {
     /// Lower without considering `tool.uv` in `pyproject.toml`, used for index and other archive
     /// dependencies.
-    pub fn from_metadata23(metadata: Metadata23) -> Self {
+    pub fn from_metadata23(metadata: MetadataResolver) -> Self {
         Self {
             metadata: Metadata::from_metadata23(metadata),
             hashes: vec![],
