+++ crates/uv-git/Cargo.toml
@@ -14,7 +14,7 @@ workspace = true
 
 [dependencies]
 cache-key = { workspace = true }
-uv-fs = { workspace = true }
+uv-fs = { workspace = true, features = ["tokio"] }
 uv-auth = { workspace = true }
 
 anyhow = { workspace = true }
